feat: User settings backend
This allows the user to set a custom email address and choose between
sending notification mails to the official one, sending them to the
custom one, or not sending them at all. It also allows setting the
locale, making the `update-user-locale` endpoint unnecessary, so it’s
deprecated now. The GET method returns some data (full name, Neptun
code etc.) that are not settable, but can be shown in the settings form
to make the displayed data more complete.