fix: Remove FL_NOCASE flag when opening ZIP
This integer constant flag is not valid for ZipArchive::open.
Furthermore, it happens to have the same value (1) as
ZipArchive::FL_CREATE, so instead of handling something in a
case-insensitive way, it actually created missing ZIP files...
They should not miss in the first place, but silently creating
them is not a great idea.